refactor: extract a class PangoLineageAliasLookup from database.cpp and the metadata store filling
