+++ include/silo/prepare_dataset.h
@@ -9,6 +9,7 @@ namespace silo {
 
 struct Partitions;
 struct PangoLineageCounts;
+class PangoLineageAliasLookup;
 
 [[maybe_unused]] void pruneSequences(
    std::istream& metadata_in,
@@ -23,7 +24,7 @@ struct PangoLineageCounts;
 );
 
 PangoLineageCounts buildPangoLineageCounts(
-   const std::unordered_map<std::string, std::string>& alias_key,
+   const PangoLineageAliasLookup& alias_key,
    std::istream& meta_in
 );
 
@@ -36,7 +37,7 @@ void partitionSequences(
    std::istream& meta_in,
    std::istream& sequence_in,
    const std::string& output_prefix,
-   const std::unordered_map<std::string, std::string>& alias_key,
+   const PangoLineageAliasLookup& alias_key,
    const std::string& metadata_file_extension,
    const std::string& sequence_file_extension
 );
